Move the res_search function to runtime package while linking in net
